RF-DETR_/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
# Roboflow Inference SDK for RF-DETR
inference>=0.13.0
pycocotools>=2.0.6
matplotlib>=3.7.1
tqdm>=4.65.0
pillow>=9.5.0
numpy>=1.24.0
